CloudSBP

Documentation for CloudSBP.

CloudSBP.BoundaryOperator

CloudSBP.BoundaryOperator

CloudSBP.CellData

CloudSBP.Dissipation

CloudSBP.Face

CloudSBP.Face

CloudSBP.LevelSetRefinery

CloudSBP.Mesh

CloudSBP.PointRefinery

CloudSBP.SBP

CloudSBP.add_face_to_boundary!

CloudSBP.add_face_to_boundary!

CloudSBP.add_face_to_boundary!

CloudSBP.apply_approx_inverse!

CloudSBP.boundary_operators

CloudSBP.boundary_sym_part

CloudSBP.boundary_sym_part

CloudSBP.boundary_sym_part

CloudSBP.build_boundary_face

CloudSBP.build_boundary_faces

CloudSBP.build_cell_stencils!

CloudSBP.build_face

CloudSBP.build_face_dissipation

CloudSBP.build_first_derivative

CloudSBP.build_interfaces

CloudSBP.build_interpolation!

CloudSBP.build_mesh

CloudSBP.build_nn_stencils!

CloudSBP.calc_error

CloudSBP.calc_moments!

CloudSBP.calc_moments!

CloudSBP.cell_null_and_part

CloudSBP.cell_quadrature

CloudSBP.cell_quadrature

CloudSBP.cell_quadrature

CloudSBP.cell_quadrature

CloudSBP.cell_quadrature_rev!

CloudSBP.cell_quadrature_rev!

CloudSBP.cell_side_rect

CloudSBP.cell_skew_part

CloudSBP.cell_symmetric_part

CloudSBP.cell_symmetric_part

CloudSBP.cell_symmetric_part

CloudSBP.cell_symmetric_part

CloudSBP.compute_sparse_constraint

CloudSBP.dgd_basis!

CloudSBP.dgd_basis_rev!

CloudSBP.diagonal_norm!

CloudSBP.diagonal_norm!

CloudSBP.diagonal_norm_rev!

CloudSBP.diff_jacobi_poly

CloudSBP.diff_proriol_poly!

CloudSBP.diff_proriol_poly!

CloudSBP.face_quadrature!

CloudSBP.get_bounding_box

CloudSBP.get_complex_step

CloudSBP.get_data

CloudSBP.get_neighbors

CloudSBP.get_null_and_part

CloudSBP.get_points_inside

CloudSBP.indices_within_parent

CloudSBP.interface_interp

CloudSBP.interface_interp

CloudSBP.interface_skew_part

CloudSBP.interface_skew_part

CloudSBP.is_center_immersed

CloudSBP.is_cut

CloudSBP.is_cut

CloudSBP.is_cut

CloudSBP.is_immersed

CloudSBP.is_immersed

CloudSBP.jacobi_poly!

CloudSBP.leaves_on_side

CloudSBP.lg_nodes

CloudSBP.lgl_nodes

CloudSBP.make_compatible!

CloudSBP.mark_cut_cells!

CloudSBP.mark_cut_faces!

CloudSBP.mass_matrix

CloudSBP.mass_obj

CloudSBP.mass_obj_grad!

CloudSBP.mass_row_sums!

CloudSBP.mass_row_sums_rev!

CloudSBP.max_leaf_stencil

CloudSBP.monomial_basis!

CloudSBP.monomial_basis_derivatives!

CloudSBP.neighbor_of_greater_or_equal_size

CloudSBP.num_leaves

CloudSBP.number_immersed

CloudSBP.obj_norm

CloudSBP.obj_norm_grad!

CloudSBP.obj_slice

CloudSBP.opt_norm!

CloudSBP.output_pyplot

CloudSBP.output_vtk

CloudSBP.penalty

CloudSBP.penalty_grad!

CloudSBP.point_data_vtk

CloudSBP.points_vtk

CloudSBP.poly_basis!

CloudSBP.poly_basis_derivatives!

CloudSBP.proriol_poly!

CloudSBP.proriol_poly!

CloudSBP.push_new_face!

CloudSBP.quadrature!

CloudSBP.refine_on_levelset!

CloudSBP.refine_on_points!

CloudSBP.set_xref_and_dx!

CloudSBP.skew_operator

CloudSBP.solve_min_norm!

CloudSBP.solve_min_norm_diff!

CloudSBP.solve_min_norm_rev!

CloudSBP.solve_norm!

CloudSBP.stencil_stats

CloudSBP.symmetric_operator

CloudSBP.tensor_basis!

CloudSBP.unsteady_vtk

RegionTrees.needs_refinement

RegionTrees.needs_refinement

RegionTrees.refine_data

RegionTrees.refine_data

CloudSBP.BoundaryOperator — Type

Data for integrals on boundaries

Each boundary condition can be associated with a BoundaryOperator, which holds data needed to compute integrals over the relevant subset of the boundary. This data is grouped by faces, since each face may require a different number of quadrature points or have a different sized stencil. If f denotes the index of some face, then xq_face[f][:,q] is the qth quadrature point on face; nrm_face[f][:,q] is the quadrature-weighted outward facing normal at xq_face[f][:,q]; dof_face[f][i] is global DOF index associated with the local interpolation index i, and; prj_face[f][q,:] is the interpolation operator from the local degrees of freedom to quadrature node q.

CloudSBP.BoundaryOperator — Method

E = BoundaryOperator(T)

Returns a BoundaryOperator with empty arrays.

CloudSBP.CellData — Type

Data associated with a RegionTree Cell.

CloudSBP.Dissipation — Type

Face-based dissipation operator

w_face are the interface areas, x_face are the face centers, R_left and R_right are interpolation operators to the faces.

CloudSBP.Face — Type

Struct for interfaces and boundary faces of cells.

CloudSBP.Face — Method

f = Face(dir, origin, widths [, data=nothing])

Construct a face with normal direction index dir, with corner at origin, and having widths dimensions.

CloudSBP.LevelSetRefinery — Type

r = LevelSetRefinery(point_subset)

Used during to refine the mesh around a given level-set. Refines until the cut cells are less than r.min_widths dimensions.

CloudSBP.Mesh — Type

Mesh data structure

CloudSBP.PointRefinery — Type

r = PointRefinery(point_subset)

Used during mesh refinement with respect to a given point cloud; see RegionTree documentation for addition information.

CloudSBP.SBP — Type

Summation-by-parts first derivative operator

S[d] holds the skew-symmetric part for direction d and E[:] holds BoundaryOperators that define the symmetric part.

CloudSBP.add_face_to_boundary! — Method

add_face_to_boundary!(bndry, cell, xc, degree, levset, levset_grad!
                      [, fit_degree=degree])

This version of the method is for faces that are the intersection between the level-set levset(x)=0 and the cell cell. In addition to the level-set, this function needs a function that can compute the gradient of the level-set, levset_grad!(g, x) which returns the gradient at x in the array g. As usual, fit_degree indicates the degree of the Bernstein polynomials used to approximate the level-set within the Algoim library.

NOTE: we use 2*degree+1 nodes in each direction for the surface quadrature, since it must integrate the normal in addition to the flux.

CloudSBP.add_face_to_boundary! — Method

add_face_to_boundary!(bndry, face, xc, degree)

Computes the quadrature points, normal vector, degrees of freedom, and interpolation operator for the face face and adds this data to the given BoundaryOperator, bndry. xc are the locations of the nodes in the stencil of face, and degree determines the order of accuracy of the quadrature (2*degree+1) and the interpolation.

CloudSBP.add_face_to_boundary! — Method

add_face_to_boundary!(bndry, face, xc, degree, levset [, fit_degree=degree])

This version of the method is for planar boundary faces that are cut by the level-set geometry defined by the function levset. The optional kwarg fit_degree indicates the degree of the Bernstein polynomials used to approximate the level-set within the Algoim library.

CloudSBP.apply_approx_inverse! — Method

apply_approx_inverse!(p, g, root, xc, dist_ref, H_tol, mu, degree, max_rank)

Finds a low-rank approximation to the inverse Hessian of the objective penalty, and applies this to -g to get a search direction p. The background mesh is root, xc are the nodes/points, dist_ref is an array of reference distances for each node, mu is the regularization parameter, degree is the target degree. The rank of the approximation is controlled by max_rank.

CloudSBP.boundary_operators — Method

E = boundary_operators(root, bc_map, bfaces, xc, levset, levset_grad!, 
                       degree [, fit_degree=degree])

Generates a set of BoundaryOperators for each of the 2*Dim planar boundary as well as the immersed surface within the domain of the root cell root. bc_map is a Dictionary that maps boundaries to boundary conditions; the key used for a planar boundaries is its side index:

1 – EAST (root.boundary.origin[1])

2 – WEST (root.boundary.origin[1] + root.boundary.width[1])

3 – SOUTH (root.boundary.origin[2])

4 – NORTH (root.boundary.origin[2] + root.boundary.width[2])

5 – BOTTOM (root.boundary.origin[3])

6 – TOP (root.boundary.origin[3] + root.boundary.width[3])

The key for the immersed surface is the string "ib". All the planar boundary faces are stored in bfaces. The DOF coordinates are given by xc. The immersed surface is defined by levset(x) = 0, and the gradient of this level set is g after calling levset_grad!(g,x). The formal polynomial accuracy of the boundary operator is degree, while fit_degree gives the degree of the Bernstein polynomial used by Algoim to fit levset.

CloudSBP.boundary_sym_part — Method

E = boundary_sym_part(face, xc, degree, levset [, fit_degree=degree])

This version of the method is for planar boundary faces that are cut by the level-set geometry defined by the function levset. The optional kwarg fit_degree indicates the degree of the Bernstein polynomials used to approximate the level-set within the Algoim library.

CloudSBP.boundary_sym_part — Method

E = boundary_sym_part(face, xc, degree)

Returns the symmetric boundary form

$\int_{\text{face}} V_i V_j n_{ ext{dir}} d\Gamma$

where the integral is over the face face with unit normal in the coordinate direction face.dir. The functions $V_i$ and $V_j$ can be regarded as degree degree basis functions at the nodes i and j within the stencil of the cell whose stencil's coordinates are xc.

CloudSBP.boundary_sym_part — Method

E = boundary_sym_part!(cell, xc, degree, levset [, fit_degree=degree])

This version of the method is for faces that are the intersection between the level-set levset(x)=0 and the cell cell. As usual, fit_degree indicates the degree of the Bernstein polynomials used to approximate the level-set within the Algoim library.

CloudSBP.build_boundary_face — Method

f = build_boundary_face(dir, cell)

Construct a boundary face with normal direction index dir having adjacent element cell. With this constructor, dir can be negative. For example, dir=1 is an East face, while dir=-1 is a West face.

CloudSBP.build_boundary_faces — Method

faces = build_boundary_faces(root)

Creates an array of boundary faces of the tree root. Boundary faces are faces of the leaves of root that touch the the East, West, North, ...etc sides of root.

CloudSBP.build_cell_stencils! — Method

build_cell_stencils!(mesh, points, degree [, tol=5*10^degree, 
                     max_iter=2*degree+1])

This method loops over the cells in mesh and constructs the stencil for each. The stencil is based on the nodes points and the polynomial degree. The function also determines the cell reference dimensions for affine scaling of the Vandermonde matrix. See build_nn_stencils! for an explanation of tol and max_iter.

NOTE: If you want a degree p SBP operator, set degree=2*p-1 so that the stencil is sufficiently large for the diagonal mass matrix.

TODO: At present this is just a front-end for build_nn_stencils!. In the future we may want an input that allows for different stencil constructions.

CloudSBP.build_face — Method

f = build_face(dir, left, right)

Construct a face with normal direction index dir having adjacent cells left and right. With this constructor, dir is always positive, and points from left to right.

CloudSBP.build_face_dissipation — Method

face_diss = build_face_dissipation(ifaces, xc, degree, levset,
                                   [, fit_degree=degree])

Returns a Dissipation type that can be used for artificial dissipation over a point cloud. iface is a list of interfaces, and xc[:,i] are the coordinates of the ith point in the cloud. degree is the polynomial degree for which the interpolations to the faces is exact. levset defines a level-set geometry, and the optional kwarg fit_degree indicates the degree of the Bernstein polynomials used to approximate the level-set within the Algoim library.

CloudSBP.build_first_derivative — Method

sbp = build_first_derivative(mesh, bc_map, xc, levset, levset_grad!, degree
                             [, fit_degree=degree])

Constructs first-derivative SBP operators of degree degree based on the background mesh, nodes xc, level-set levset (and its gradient, levset_grad!). fit_degree sets the polynomial degree used by Algoim to approximate the level set. bc_map determines the type of each boundary.

PRE: The cell norms must be stored in the cell data.wts[:] attribute.

CloudSBP.build_interfaces — Method

faces = build_interfaces(root)

Creates an array of faces between adjacent leaves (i.e. cells) of the tree root. This array does not include boundary faces that have only one adjacent cell; see build_boundary_faces for such a list.

CloudSBP.build_interpolation! — Method

build_interpolation!(interp, degree, x, x_interp, xref, dx)

Creates the matrix interp that performs interpolation from nodes x to points x_interp, which is degree total polynomial degree exact. The arrays xref and dx are used to shift and scale the coordinates to improve the conditioning of the Vandermonde matrix.

Note: This routine assumes that x is unisolvent. Also, the matrix interp is the minmum-norm solution when there are more nodes in x than total degree basis functions.

CloudSBP.build_mesh — Method

mesh = build_mesh(points, widths, levset, min_width 
                  [, origin=SVector(ntuple(i -> 0.0, Dim))])

Builds the background Cartesian mesh. The root cell has its origin at origin and has widths lengths. The level-set levset defines any immersed boundary based on where levset(x) = 0; the domain is the intersection of the root cell's domain and where levset(x) >= 0. The mesh is refined based on the given points points and it is refined at the zero level-set until the cut cell have dimensions of min_width or smaller.

NOTE: If you want a degree p SBP operator, set degree=2*p-1 so that the stencil is sufficiently large for the diagonal mass matrix.

CloudSBP.build_nn_stencils! — Method

build_nn_stencils!(root, points, degree [, tol=5*10^degree, 
                   max_iter=2*degree+1])

The stencil for each leaf in the tree root is determined and the indices of the stencil are stored in the leaves. The stencil is determined using k neareast neighbors, where k is the number of points needed for a well-conditioned Vandermonde matrix of total degree degree. The tolerance tol is used to determine what is considered well-conditioned. A maximum of max_iter iterations are used to find a suitable stencil; after the maximum iterations are exceeded, the method accepts the stencil as is.

CloudSBP.calc_error — Method

L2_error, max_error = calc_error(exact, root, xc, degree, u, levset [, 
                                 quad_degree=2*degree, fit_degree=degree])

Returns the L2 and max error of a given solution u. The exact solution is provided by the function exact. The background mesh is root and the nodes are xc. degree is the degree of the solution, while quad_degree and fit_degree are the quadrature and level-set levset fitting degrees, respectively.

CloudSBP.calc_moments! — Method

m = calc_moments!(root, levset, degree, fit_degree)

Returns the first total degree integral moments for all cells in the tree defined by root. The tree must have been preprocessed to identify poentially cut and immersed cells using the levset level-set function. In addition, the cell.data.xref and cell.data.dx fields must contain the reference origin for each cell.

WARNING: The signature of the function levset must of the form levset(Vector{Float64})::Float64, because this assumption is used when it is wrapped using csafe_function.

CloudSBP.calc_moments! — Method

m = calc_moments!(root, degree)

This variant is useful when you want moments for all cells in the tree root, i.e. no level-set function.

CloudSBP.cell_null_and_part — Method

w = cell_null_and_part(degree, xc, moments, xref, dx, Val(Dim))

Given a set of total degree polynomial moments, computes a quadrature that is exact for those moments based on the nodes xc as well as the null space of the quadrature conditions The arrays xref and dx are used to shift and scale, respectively, the nodes in xc to improve conditioning of the Vandermonde matrix. The same scaling and shifts must have been applied when computing the integral moments.

CloudSBP.cell_quadrature — Method

w = cell_quadrature(degree, xc, moments, xref, dx, Val(Dim))

Partially complexified version of cell_quadrature; partially, because solve_min_norm! does not work with complex variables, so that function is differentiated explicitly here.

CloudSBP.cell_quadrature — Method

w = cell_quadrature(degree, xc, xq, wq, Val(Dim))

Complexified version of cell_quadrature for testing derivatives.

CloudSBP.cell_quadrature — Method

w = cell_quadrature(degree, xc, moments, xref, dx, Val(Dim))

Given a set of total degree polynomial moments, computes a quadrature that is exact for those moments based on the nodes xc. The arrays xref and dx are used to shift and scale, respectively, the nodes in xc to improve conditioning of the Vandermonde matrix. The same scaling and shifts must have been applied when computing the integral moments.

CloudSBP.cell_quadrature — Method

w = cell_quadrature(degree, xc, xq, wq, Val(Dim))

Given a quadrature rule (xq,wq) that is exact for polynomials of degree degree, computes a new quadrature for the same domain based on the nodes xc. This version computes the moments from scratch.

CloudSBP.cell_quadrature_rev! — Method

cell_quadrature_rev!(xc_bar, degree, xc, moments, xref, dx, w_bar, Val(Dim))

Reverse mode differentiated of the moment-based variant of cell_quadrature. Returns the derivatives of dot(w, w_bar) with respect to xc in the array xc_bar. All other inputs are the same as the moment-based variant of cell_quadrature.

CloudSBP.cell_quadrature_rev! — Method

cell_quadrature_rev!(xc_bar, degree, xc, xq, wq, w_bar, Val(Dim))

Reverse mode differentiated cell_quadrature. Returns the derivatives of dot(w, w_bar) with respect to xc in the array xc_bar. All other inputs are the same as cell_quadrature.

CloudSBP.cell_side_rect — Method

rect = build_cell_face(dir, cell)

Construct a HyperRectangle for given cell on side dir.

CloudSBP.cell_skew_part — Method

S = cell_skew_part(E, H, cell, xc, degree)

Returns the skew-symmetric parts of SBP diagonal-norm operators for the element cell based on the nodes xc. The operator is exact for polynomials of total degree degree. The diagonal norm for the cell is provided in the array H, which must be exact for degree 2*degree - 1 polynomials over xc. Finally, the symmetric part of the SBP operators must be provided in E. Note that E and the returned S are three dimensional arrays, with E[:,:,d] and S[:,:,d] holding the operators for the direction d.

NOTE: E and H must be compatible, in the sense of SBP operators.

CloudSBP.cell_symmetric_part — Method

E = cell_symmetric_part(cell, xc, degree, levset [,fit_degree=degree])

Returns the symmetric part of the first-derivative SBP operator for the possibly cut cell cell. The point cloud associated with cell is xc, and the boundary operator uses 2*degree exact quadrature. The fit_degree input indicates the degree of the Bernstein polynomial used by Algoim to approximate the level set.

CloudSBP.cell_symmetric_part — Method

E = cell_symmetric_part(cell, xc, degree)

Returns the symmetric part of the first-derivative SBP operator for the uncut cell cell. The point cloud associated with cell is xc, and the boundary operator is 2*degree exact for boundary integrals.

Note: This version recomputes the 1D quadrature rule each time, and involves several allocations.

CloudSBP.cell_symmetric_part — Method

E = cell_symmetric_part(cell, faces, xc, degree, levset [,fit_degree=degree])

Returns the symmetric part of the first-derivative SBP operator for the possibly cut cell cell. The point cloud associated with cell is xc, and the boundary operator uses 2*degree exact quadrature. The fit_degree input indicates the degree of the Bernstein polynomial used by Algoim to approximate the level set. This variant loops over the faces of the cell, which is not necessary for uncut cells but it is necessary for cut cells.

CloudSBP.cell_symmetric_part — Method

E = cell_symmetric_part(cell, faces, xc, degree)

Returns the symmetric part of the first-derivative SBP operator for the uncut cell cell. The point cloud associated with cell is xc, and the boundary operator is 2*degree exact for boundary integrals. This variant loops over the faces of the cell, which is not necessary for uncut cells but it is necessary for cut cells.

Note: This version recomputes the 1D quadrature rule each time, and involves several allocations.

CloudSBP.compute_sparse_constraint — Method

A, b = compute_sparse_constraint(root, wp, Z, H_tol)

Returns the sparse matrix A and vector b that make up the linear inequality for the diagonal mass matrix. root is used to loop over the cells and assemble the null-space matrices stored in Z and the minimum-norm quadrature stored in wp. H_tol is an array of tolerances for the diagonal entries.

CloudSBP.dgd_basis! — Method

dgd_basis!(phi, degree, xc, xq, work, Val(Dim))

Evaluates the DGD basis functions phi at the points xq. The basis functions are of total degree degree, and they are defined by the centers xc. The array work is used for temporary storage, and the value type Dim is used to dispatch on the relevant dimension.

If ndims(phi) > 2, it is assumed that both the basis functions (stored in phi [:,:,q]) and their derivatives (stored in phi[:,:,2:Dim+1]) are to be computed. Otherwise, the basis functions are computed and returned in the 2d array phi[:,:].

CloudSBP.dgd_basis_rev! — Method

dgd_basis_rev!(xc_bar, phi_bar, degree, xc, xq, xref, dx, Val(Dim))

Reverse mode algorithmic differentiation of dgd_basis!. The output xc_bar is the derivative of the output (defined implicitly by the input phi_bar) as a function of the coordinates in xc. As in dgd_basis!, the basis functions are of total degree degree, evaluated at points xq, and they are defined by the centers xc. The value type Dim is used to dispatch on the relevant dimension.

Note: Only differentiates the version of dgd_basis! that returns the basis functions, not their derivatives.

CloudSBP.diagonal_norm! — Method

diagonal_norm!(H, root)

Assembles the diagonal norm H using the particular cell-based norm stored in the cells' data.wts field. root stores the mesh.

CloudSBP.diagonal_norm! — Method

diagonal_norm!(H, root, points, degree)

Constructs a diagonal norm (i.e. quadrature) for the nodes points and using the background mesh defined by root.

PRE: The cells in root must have their data.moments, data.xref, and data.dx fields set.

CloudSBP.diagonal_norm_rev! — Method

diagonal_norm_rev!(points_bar, H_bar, root, points, degree)

Reverse-mode differentiation of the weighted norm, dot(H, H_bar) with respect to the node coordinates in points.

PRE: The cells in root must have their data.moments, data.xref, and data.dx fields set.

CloudSBP.diff_jacobi_poly — Method

dp = diff_jacobi_poly(x, alpha, beta, N)

Evaluate the first derivative of a Jacobi Polynomial at the points x. See the companion code for jacobipoly for an explanation of the other parameters.

CloudSBP.diff_proriol_poly! — Method

diff_proriol_poly!(dP, x, y, z, i, j, k)

Evaluate the derivatives of a Proriol orthogonal polynomial basis function on the right tetrahedron. The arrays x,y, and z are the locations at which to evaluate the polynomial, and the integers i,j, and k define the basis function to evaluate. The result is stored in dP.

Notes: the derivatives are computed using the complex-step method (since there are many outputs and only 3 inputs); therefore, a different method should be used for verification of this method.

CloudSBP.diff_proriol_poly! — Method

diff_proriol_poly!(dP, x, y, i, j)

Evaluate the derivatives of a Proriol orthogonal polynomial basis function on the right triangle. The arrays x and y give the locations at which to evaluate the polynomial, and the integers i and j define the basis function to evaluate. The result is stored in dP.

CloudSBP.face_quadrature! — Method

face_quadrature!(xq, wq, rect, x1d, w1d, dir)

Builds a tensor-product quadrature rule on a Dim-1 hyperrectangle. The quadrature is returned in the updated arrays xq and wq, which hold the locations and weights, respectively. The rule is based on the one-dimensional quadrature defined by x1d and w1d. The Dim-1 hyperrectuangle has a boundary defined by rect, which has a unit normal in the Cartesian index dir. Note that rect is defined in Dim dimensions, and has rect.widths [dir] = 0, i.e. it has zero width in the dir direction.

CloudSBP.get_bounding_box — Method

box_center, box_dx = get_bounding_box(rect, x)

For a given HyperRectangle rect and point cloud x, returns the bounding box center and lengths that enclose both rect and x.

CloudSBP.get_complex_step — Method

get_complex_step(T)

returns an appropriate complex-step size for the given type

CloudSBP.get_data — Method

data = get_data(cell, child_indices)

Returns a CellData struct based on the given cell.

CloudSBP.get_neighbors — Method

neighbors = get_neighbors(cell, side)

Constructs a vector of all neighbors on the side boundary of leaf cell.

Based on the paper here: http://www.cs.umd.edu/~hjs/pubs/SameCVGIP89.pdf See also https://geidav.wordpress.com/2017/12/02/advanced-octrees-4-finding-neighbor-nodes/

CloudSBP.get_null_and_part — Method

wp, Z, num_vars = get_null_and_part(root, xc, degree)

Returns a vector of nullspaces, Z, and particular solutions, wp, for each cell's quadrature problem. Also returns the total number of degrees of freedom. Note that wp and Z are Vectors of Vector and Matrix, respectively.

PRE: The cells in root must have their data.moments, data.xref, and data.dx fields set.

CloudSBP.get_points_inside — Function

inside = get_points_inside(rect, points [, indices=[1:size(points,2)]])

Returns the subset of integers from indices corresponding to points from points that are inside the hyperrectangle rect.

CloudSBP.indices_within_parent — Method

I = indices_within_parent(cell)

Returns the Cartesian indices of cell with respect to its parent. This function assumes cell has a parent and will fail if it does not.

CloudSBP.interface_interp — Method

Rl, Rr, x, w  = interface_interp(face, xc_left, xc_right, degree)

Returns interpolation operators Rl and Rr from xc_left and xc_right to x, respectively, where x is the returned averaged center of the face face. w is the weight associate with face (its area).

CloudSBP.interface_interp — Method

Rl, Rr, x, w  = interface_interp(face, xc_left, xc_right, degree, levset
                                 [, fit_degree=degree])

This version of the method is for planar interfaces that are cut by a level-set geometry defined by the function levset. The optional kwarg fit_degree indicates the degree of the Bernstein polynomials used to approximate the level-set within the Algoim library.

CloudSBP.interface_skew_part — Method

Sface = interface_skew_part(face, xc_left, xc_right, degree)

Constructs the form

$\int_{\text{face}} V_i V_j d\Gamma$

where the integral is over the face face with unit normal in the coordinate direction face.dir. The functions $V_i$ and $V_j$ can be regarded as degree degree basis functions at the nodes i and j within the stencil of the left and right cells, respectively.

CloudSBP.interface_skew_part — Method

Sface = interface_skew_part(face, xc_left, xc_right, degree, levset
                            [, fit_degree=degree])

This version of the method is for planar interfaces that are cut by a level-set geometry defined by the function levset. The optional kwarg fit_degree indicates the degree of the Bernstein polynomials used to approximate the level-set within the Algoim library.

CloudSBP.is_center_immersed — Method

inside = is_center_immersed(rect, levset)

Returns true if the center of rect is inside the level-set levset, that is, phi(xc) < 0.0.

CloudSBP.is_cut — Method

cut = is_cut(cell)

Returns false if cell is not cut; note that a return of true, however, only indicates that the cell may be cut.

CloudSBP.is_cut — Method

cut = is_cut(face)

Returns false if face is not cut; note that a return of true, however, only indicates that the face may be cut.

CloudSBP.is_cut — Method

cut = is_cut(rect, levset [, fit_degree=2])

Returns true if the HyperRectangle rect is intersected by the level-set levset, and returns false otherwise.

CloudSBP.is_immersed — Method

im = is_immersed(cell)

Returns true if cell is not cut and its center is immersed.

CloudSBP.is_immersed — Method

im = is_immersed(face)

Returns true if face is not cut and its center is immersed.

CloudSBP.jacobi_poly! — Method

jacobi_poly!(p, x, alpha, beta, N, work)

Evaluate Jacobi polynomial at the points x and return in p. Based on JacobiP in Hesthaven and Warburton's nodal DG book. alpha and beta are parameters that define the type of Jacobi Polynomial (alpha + beta != 1). N is the polynomial degree, not the number of nodes. The array work should be twice the length of x.

CloudSBP.leaves_on_side — Method

leaves = leaves_on_side(cell, side)

Returns a Vector of leaves on boundary index side of given root cell. If cell is nothing, the Vector is returned empty.

CloudSBP.lg_nodes — Function

x, w = lg_nodes(N [, T=Float64])

Computes the Legendre-Gauss (LG) quadrature nodes x and weights w on the interval [-1,1]. The LG nodes are the zeros of PN(x), where PN(x) denotes the Nth Legendre polynomial.

Julia version adapted from Matlab code written by Greg von Winckel - 02/25/2004 Contact: gregvw@chtm.unm.edu

CloudSBP.lgl_nodes — Function

x, w = lgl_nodes(N [, T=Float64])

Computes the Legendre-Gauss-Lobatto (LGL) quadrature nodes x and weights w on the interval [-1,1]. The LGL nodes are the zeros of (1-x^2)*P'N(x), where PN(x) denotes the Nth Legendre polynomial.

Julia version adapted from Matlab code written by Greg von Winckel - 04/17/2004 Contact: gregvw@chtm.unm.edu

CloudSBP.make_compatible! — Method

make_compatible!(E, H, cell, xc, degree)

Modifies the SBP symmetric operator E for cell cell such that it is compatible with the diagonal norm H. The operators are defined over the nodes xc and are for a degree degree exact SBP first-derivative operator.

CloudSBP.mark_cut_cells! — Method

mark_cut_cells!(root, levset)

Identifies cells in the tree root that may be cut be the level-set levset.

CloudSBP.mark_cut_faces! — Method

mark_cut_faces!(faces, levset)

Identifies faces in the list faces that may be cut be the level-set levset.

CloudSBP.mass_matrix — Method

M = mass_matrix(root, xc, degree)

Returns the DGD mass matrix for degree degree based on the mesh in the tree root and the centers in xc.

CloudSBP.mass_obj — Method

obj = mass_obj(root, xc, xc_init, dist_ref, H_tol, mu, degree)

Compute the objective that seeks to minimize the change in the node locations while ensuring a positive lumped mass matrix. root is the mesh, xc are the nodes being varied, and xc_init are the initial node locations. dist_ref are reference lengths, and H_tol is an array of tolerances for the lumped mass value at each node; that is, sum_{j} M[i,j] >= H_tol[i] for the constraint to be satisfied. Finally, mu scales the regularization term, and degree is the target exactness of the rule.

CloudSBP.mass_obj_grad! — Method

mass_obj_grad!(g, root, xc, xc_init, dist_ref, H_tol, mu, degree)

Computes the derivative of mass_obj with respect to xc. See mass_obj for and explanation of the remaining parameters.

CloudSBP.mass_row_sums! — Method

mass_row_sums!(lumped_mass, root, xc, degree)

Fills the array lumped_mass with the sum of the rows (or columns) of the symmetric DGD mass matrix of degree degree based on the cloud xc.

CloudSBP.mass_row_sums_rev! — Method

mass_row_sums_rev!(xc_bar, lumped_mass_bar, root, xc, degree)

Reverse-mode differentiation of the weighted norm, dot(lumped_mass, lumped_mass_bar) with respect to the node coordinates in points.

CloudSBP.max_leaf_stencil — Method

max_stencil = max_leaf_stencil(root)

Returns the largest stencil (i.e. number of points used for DGD basis) over all leaves in the mesh defined by root.

CloudSBP.monomial_basis! — Method

monomial_basis!(V, degree, x, Val(N))

This method computes the monomial basis of total degree degree at the points x. The type parameter N is used to select the appropriate spatial dimension. These methods are useful for testing, but given the ill-conditioning of the corresponding Vandermonde matrix, they should not be used to construct the DGD basis functions.

CloudSBP.monomial_basis_derivatives! — Method

monomial_basis_derivatives!(dV, degree, x, Val(N))

This method and its variants compute the first derivatives of monomial basis of total degree degree at the points x. The type parameter N is used to select the appropriate spatial dimension.

CloudSBP.neighbor_of_greater_or_equal_size — Method

nbr = neighbor_of_greater_or_equal_size(cell, side)

Returns the cell on boundary index side of cell that is of the same size or larger than cell. Returns nothing` if no such neighbor exists.

CloudSBP.num_leaves — Method

n = num_leaves(root)

Returns the total number of leaves in the given root.

CloudSBP.number_immersed — Method

count = number_immersed(faces)

Returns the number of faces that are definitely immersed. Note that this is a lower bound, since the immeresed check is conservative.

CloudSBP.obj_norm — Method

obj = obj_norm(root, wp, Z, y, rho, num_nodes)

Compute the objective to maximize the minimum norm. root is the mesh, which is mostly needed to known the stencil for each cell. wp is a vector of vectors holding the particular solution for each cell, and Z is the nullspace for each cell's problem.

NOTE: this objective is deprecated.

CloudSBP.obj_norm_grad! — Method

obj_norm_grad!(g, root, wp, Z, y, num_nodes)

Compute the gradient of the objective to maximize the minimum norm. root is the mesh, which is mostly needed to known the stencil for each cell. wp is a vector of vectors holding the particular solution for each cell, and Z is the nullspace for each cell's problem.

CloudSBP.obj_slice — Method

obj_slice(root, xc, degree, H_tol, mu, dist_ref, node, pert)

Returns an array of objective function values along a slice based on perturbing node node in the direction pert. Used for visualizing.

CloudSBP.opt_norm! — Method

H = opt_norm!(root, xc, degree, H_tol, mu, dist_ref, max_rank)

Finds an approximate minimizer for the objective penalty with respect to the node coordinates xc and based on the background mesh root. Once an approximate solution is found, the corresponding H norm is returned; note that the xc coordinates are altered in the process. The parameter max_rank determines the rank of the approximate Jacobian of the norm with respective to the nodes, and this approximate Jacobian is used to form an approximate inverse Hessian. See penalty for explanations of the other parameters.

CloudSBP.output_pyplot — Method

xplot, uplot = output_pyplot(root, xc, degree, u [, num_pts=degree+1])

Outputs "meshgrid" type data for each cell in root for plotting the contours of the state u using Matplotlib. The solution is of degree degree and num_pts samples are used in each direction on each cell.

CloudSBP.output_vtk — Method

vtk = output_vtk(root, xc, degree, u [, num_pts=degree+1,
                 filename="solution", save=true])

Creates a VTK file for a scalar solution based on the mesh in root, the points xc, and the solution coefficients in u. The reconstruction used is of degree degree, although only the solution is only interpolated to the cell vertices at this time (i.e. the cell's themselves are not high-order in Paraview).

CloudSBP.penalty — Method

obj = penalty(root, xc, xc_init, dist_ref, H_tol, mu, degree)

Compute the objective that seeks to minimize the change in the node locations while ensuring positive quadrature weights. root is the mesh, xc are the nodes being varied, and xc_init are the initial node locations. dist_ref are reference lengths, and H_tol is an array of tolerances for the quadrature weight at each node; that is, H[i] >= H_tol[i] for the constraint to be satisfied. Finally, mu scales the regularization term, and degree is the target exactness of the rule.

CloudSBP.penalty_grad! — Method

penalty_grad!(g, root, xc, xc_init, dist_ref, H_tol, mu, degree)

Computes the derivative of penalty with respect to xc. See penalty for an explanation of the remaining parameters.

CloudSBP.point_data_vtk — Method

point_data_vtk(xc, data, label [, filename="point_data"])

Writes the points xc to the file "point_data". data is a (possibly empty) Vector of arrays that holds data to be associated with each node, and label is a label associated with it.

Example

# write data to plot the norm and its sign at each node
point_data_vtk(xc, [abs.(H), sign.(H)], ["norm-mag", "norm-sign"])

CloudSBP.points_vtk — Method

points_vtk(xc [, filename="points"])

Writes the cloud points xc to the file "points.vtu".

CloudSBP.poly_basis! — Method

poly_basis!(V, degree, x, work, Val(N))

This method and its variants compute the approporiate polynomial basis of total degree degree at the points x. The type parameter N is used to select the appropriate spatial dimension. The size of work should be N+1 times the number of rows in V.

CloudSBP.poly_basis_derivatives! — Method

poly_basis_derivatives!(dV, degree, x, Val(N))

This method and its variants compute the first derivatives of polynomial basis of total degree degree at the points x. The type parameter N is used to select the appropriate spatial dimension.

CloudSBP.proriol_poly! — Method

proriol_poly!(p, x, y, z, i, j, k, work)

Evaluate a Proriol orthogonal polynomial basis function on the right tetrahedron and return in array p. The arrays x,y, and z are the locations at which to evaluate the polynomial, and the integers i,j, and k define the basis function to evaluate; see Hesthaven and Warburton's Nodal DG book, for example, for a reference. The work array should be at least 4 times the length of p.

CloudSBP.proriol_poly! — Method

proriol_poly!(p, x, y, i, j, work)

Evaluate a Proriol orthogonal polynomial basis function on the right triangle and return in array p. The arrays x and y give the locations at which to evaluate the polynomial, and the integers i and j define the basis function to evaluate; see Hesthaven and Warburton's Nodal DG book, for example, for a reference. work is a vector that should be 3 times the length of p.

CloudSBP.push_new_face! — Method

xq, nrm, dof, prj = push_new_face!(bndry, Dim, num_nodes, num_quad)

Creates new memory at the end of the fields in the given BoundaryOperator and returns references to the newly created arrays.

CloudSBP.quadrature! — Method

quadrature!(xq, wq, rect, x1d, w1d)

Builds a tensor-product quadrature for the domain defined by rect using the one-dimensional quadratrure defined by the nodes x1d and the weights w1d. The tensor-product quadrature nodes and weights are given by xq and wq, respectively. These are stored in 2-d and 1-d array formats.

CloudSBP.refine_on_levelset! — Method

refine_on_levelset!(root, points, levset, min_widths)

Refines the leaves in root until each cut leaf is less than min_widths dimensions. The array of points are needed so that refined leaves have the necessary data.

CloudSBP.refine_on_points! — Method

refine_on_points!(root, points)

Refines the tree root until each cell has at most one of the points in points.

CloudSBP.set_xref_and_dx! — Method

set_xref_and_dx!(root, points)

For each leaf in root, finds a reference position and reference scale for leaf.data.xref and leaf.data.dx, respectively. The reference position is, roughly, the average of the coordinates points[:,leaf.data.points] and the leaf's boundary. The reference scale is the extent of the coordinates and boundary.

CloudSBP.skew_operator — Method

S = skew_operator(root, ifaces, bfaces, xc, levset, degree
                  [, fit_degree=degree, use_cell_wts=true])

Constructs the skew-symmetric part of a (global), first-derivative SBP operator. The integration mesh is given by root and xc defines the cloud of distributed nodes where the degrees of freedom are stored. ifaces is an array of interfaces and bfaces is an array of boundary faces corresponding to root. levset is a function that defines the immersed geomtry, if any. The skew-symmetric matrix is degree degree exact. Finally, fit_degree gives the polynomial degree of the Bernstein polynomials used to approximate levset by the Algoim library. If use_cell_wts is true, the cell-based norm is formed using the data in cell.data.wts.

CloudSBP.solve_min_norm! — Method

solve_min_norm!(w, V, b)

Finds the minimum norm solution to transpose(V) w = b.

CloudSBP.solve_min_norm_diff! — Method

solve_min_norm_diff!(w, dw, V, dV, b)

Forward mode of solve_min_norm! treating b as constant. This is useful for verifying the reverse mode implementation.

CloudSBP.solve_min_norm_rev! — Method

solve_min_norm_rev!(V_bar, w_bar, V, b)

Reverse mode of solve_min_norm!. On entry, w_bar holds the derivatives of the objective w.r.t. the weights. On exit, V_bar holds the derivatives of the objective w.r.t. the matrix V.

CloudSBP.solve_norm! — Method

H, success = solve_norm!(root, xc, degree, H_tol [, verbose=false])

Attempts to solve the norm-inequality problem using the interior-point method implemented in the Tulip.jl package. root defines the background mesh, xc are the nodes, degree is the target degree of the norm, and H_tol holds the lower bound on the norm entries. If verbose is true, the "OutputLevel" attribute of Tulip is set to 1 (so more output is provided). The method returns the diagonal norm, H, and a Bool, success, that indicates whether Tulip was successful (success=true) or not. If the problem was not successfully solved, H is returned with the minimum-norm quadrature weights (which likely have some negative weights).

PRE: The cells in root must have their data.moments, data.xref, and data.dx fields set.

CloudSBP.stencil_stats — Method

max_stencil, avg_stencil = stencil_stats(mesh)

Returns the maximum cell stencil size and the average stencil size. Only cells that are non immersed are included, since immersed cells should have empty stencils.

PRE: The cells in mesh must have their stencils defined; this should be the case if a high-level mesh construction was used (e.g. build_mesh).

CloudSBP.symmetric_operator — Method

E = symmetric_operator(root, ifaces, bfaces, xc, levset, degree
                       [, fit_degree=degree])

Constructs the symmetric part of a (global), first-derivative SBP operators as an SVector of sparse matrices; only the symmetric part is stored. The integration mesh is given by root and xc defines the cloud of distributed nodes where the degrees of freedom are stored. ifaces is an array of intefaces and bfaces is an array of boundary faces corresponding to root. levset is a function that defines the immersed geomtry, if any. The symmetric matrix is degree degree exact. Finally, fit_degree gives the polynomial degree of the Bernstein polynomials used to approximate levset by the Algoim library.

Note: The sparse matrices are not useful for much other than verifying the accuracy of the global skew-symmetric parts; this is because we do not distinguish different boundaries, nor do we provide the decomposition of the symmetric part into interpolation operators and face quadrature.

CloudSBP.tensor_basis! — Method

tensor_basis!(V, degree, x, Val(N))

This method computes the tensor-product basis of degree degree at the points x. The type parameter N is used to select the appropriate spatial dimension.

CloudSBP.unsteady_vtk — Method

unsteady_vtk(root, xc, degree, t, sol [, num_pts=degree+1,
             filename="unsteady"])

Writes a Paraview collection in order to visualizae an unsteady solution. The solution at time t[k] is stored in sol[:,k]. See output_vtk for an explanation of the other inputs.

RegionTrees.needs_refinement — Method

This method is for determining if cut cells need further refinement.

RegionTrees.needs_refinement — Method

refine = needs_refinement(r, cell)

Returns true if cell has more than one point in it.

RegionTrees.refine_data — Method

This method, which is almost identical to the above method, is used when refining cells that are cut by a given levelset.

RegionTrees.refine_data — Method

child_data = refine_data(r, cell, indices)

Returns data for child of cell with indices.